pyproject.toml
[project]
name = "devils-advocate-cli"
version = "0.1.0"
description = "A standalone CLI for Devil's Advocate workflow"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer[all]",
    "litellm",
    "python-dotenv",
    "rich",
    "mcp",
]

[project.scripts]
devils-advocate = "devils_advocate_cli.main:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"devils_advocate_cli.data" = ["*.md"]
